+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.1.0-wip
 
 - Initial version.
+- Expose `Mutex` and `ConditionVariable`
+- Implement `Mailbox`.